Simplify text wrapping behavior by going linewise
Instead of heuristically detecting quotes and bulleted lists, just wrap
text line by line. GitHub issues and pull requests are usually not
wrapped at all, in which case there is no need to look at neighboring
lines. Replies left via email or the GitHub command line are probably
wrapped to 72 characters, so there should be no change with -w72. This
will cause some erratic behavior in the case that text is wrapped to
a different width than that given to -w, but the user is welcome to
reflow the text.